add attribution for server/auth/basic_auth.rs
