+++ rust/oas-core/src/server/auth/mod.rs
@@ -14,11 +14,13 @@ use time::{Duration, OffsetDateTime};
 // use schemars::JsonSchema;
 // use serde::Serialize;
 
+mod basic_auth;
 mod password;
 mod sessions;
 mod store;
 mod structs;
 
+use basic_auth::BasicAuth;
 pub use sessions::{SessionInfo, Sessions};
 use store::UserStore;
 use structs::{LoginRequest, LoginResponse, RegisterRequest};
@@ -351,72 +353,4 @@ fn generate_session_id() -> String {
     let random_bytes: [u8; 32] = rand::random();
     let encoded = base32::encode(base32::Alphabet::Crockford, &random_bytes[..]);
     encoded.to_lowercase()
-}
-
-#[derive(Debug)]
-pub struct BasicAuth {
-    /// Required username
-    pub username: String,
-
-    /// Required password
-    pub password: String,
-}
-
-impl BasicAuth {
-    /// Creates a new [BasicAuth] struct/request guard from a given plaintext
-    /// http auth header or returns a [Option::None] if invalid
-    pub fn new<T: Into<String>>(auth_header: T) -> Option<Self> {
-        let key = auth_header.into();
-
-        if key.len() < 7 || &key[..6] != "Basic " {
-            return None;
-        }
-
-        let (username, password) = decode_basic_auth(&key[6..])?;
-
-        Some(Self { username, password })
-    }
-}
-
-#[rocket::async_trait]
-impl<'r> FromRequest<'r> for BasicAuth {
-    type Error = ();
-
-    async fn from_request(request: &'r Request<'_>) -> Outcome<Self, Self::Error> {
-        let keys: Vec<_> = request.headers().get("Authorization").collect();
-        match keys.len() {
-            0 => Outcome::Forward(()),
-            1 => match BasicAuth::new(keys[0]) {
-                Some(auth_header) => Outcome::Success(auth_header),
-                None => Outcome::Failure((Status::BadRequest, ())),
-            },
-            _ => Outcome::Failure((Status::BadRequest, ())),
-        }
-    }
-}
-
-impl From<BasicAuth> for LoginRequest {
-    fn from(auth: BasicAuth) -> Self {
-        Self {
-            username: auth.username,
-            password: auth.password,
-        }
-    }
-}
-
-/// Decodes a base64-encoded string into a tuple of `(username, password)` or a
-/// [Option::None] if badly formatted, e.g. if an error occurs
-fn decode_basic_auth<T: Into<String>>(base64_encoded: T) -> Option<(String, String)> {
-    let decoded_creds = match base64::decode(base64_encoded.into()) {
-        Ok(vecu8_creds) => String::from_utf8(vecu8_creds).unwrap(),
-        Err(_) => return None,
-    };
-
-    let split_vec: Vec<&str> = decoded_creds.splitn(2, ":").collect();
-
-    if split_vec.len() < 2 {
-        None
-    } else {
-        Some((split_vec[0].to_string(), split_vec[1].to_string()))
-    }
 }
